Moved Configuration to respective package
